Move Target into mod.rs
