+++ crates/builder/src/graph/csr.rs
@@ -15,6 +15,7 @@ use std::{
 use rayon::prelude::*;
 
 use crate::compat::*;
+
 use crate::{
     graph_ops::{DeserializeGraphOp, SerializeGraphOp, ToUndirectedOp},
     index::Idx,
@@ -22,8 +23,8 @@ use crate::{
     DirectedDegrees, DirectedNeighbors, DirectedNeighborsWithValues, Error, Graph,
     NodeValues as NodeValuesTrait, SharedMut, UndirectedDegrees, UndirectedNeighbors,
     UndirectedNeighborsWithValues,
+    Target,
 };
-
 #[cfg(feature = "dotgraph")]
 use crate::input::DotGraph;
 #[cfg(feature = "dotgraph")]
@@ -58,40 +59,6 @@ pub enum CsrLayout {
 pub struct Csr<Index: Idx, NI, EV> {
     offsets: Box<[Index]>,
     targets: Box<[Target<NI, EV>]>,
-}
-
-/// Represents the target of an edge and its associated value.
-#[derive(Clone, Copy, Debug)]
-#[repr(C)]
-pub struct Target<NI, EV> {
-    pub target: NI,
-    pub value: EV,
-}
-
-impl<T: Ord, V> Ord for Target<T, V> {
-    fn cmp(&self, other: &Self) -> std::cmp::Ordering {
-        self.target.cmp(&other.target)
-    }
-}
-
-impl<T: PartialOrd, V> PartialOrd for Target<T, V> {
-    fn partial_cmp(&self, other: &Self) -> Option<std::cmp::Ordering> {
-        self.target.partial_cmp(&other.target)
-    }
-}
-
-impl<T: PartialEq, V> PartialEq for Target<T, V> {
-    fn eq(&self, other: &Self) -> bool {
-        self.target.eq(&other.target)
-    }
-}
-
-impl<T: Eq, V> Eq for Target<T, V> {}
-
-impl<T, EV> Target<T, EV> {
-    pub fn new(target: T, value: EV) -> Self {
-        Self { target, value }
-    }
 }
 
 impl<Index: Idx, NI, EV> Csr<Index, NI, EV> {
