+++ crates/builder/src/prelude.rs
@@ -1,8 +1,8 @@
 pub use crate::builder::GraphBuilder;
 
+pub use crate::graph::Target;
 pub use crate::graph::csr::CsrLayout;
 pub use crate::graph::csr::DirectedCsrGraph;
-pub use crate::graph::csr::Target;
 pub use crate::graph::csr::UndirectedCsrGraph;
 
 pub use crate::graph_ops::DegreePartitionOp;
